feature: add SingleImageProvider
Sometimes it's handy to be able to use a single image as a layer (mainly want you
want to quick test without the hassle of setting up a wms server for instance).

This provider is really simple, and just compute the offset in the image to use
to match the tile extent.

Example usage:
    view.addLayer({
        url: 'http://localhost:8080/foo.png',
        networkOptions: { crossOrigin: 'anonymous' },
        type: 'color',
        protocol: 'image',
        id: 'myimage',
    });